Make the slide page with memoize for the page should not be rendering with content no change.
